💄IAlertService - add interface
