+++ Satori.AppServices/Services/Abstractions/IAlertService.cs
@@ -1,0 +1,11 @@
+﻿using Satori.AppServices.ViewModels.AlertServices;
+
+namespace Satori.AppServices.Services.Abstractions;
+
+public interface IAlertService
+{
+    void BroadcastAlert(Exception ex);
+    void BroadcastAlert(string message);
+    void ClearAlert();
+    AlertViewModel Subscribe();
+}
