+++ Satori/Pages/Components/AlertBanner.razor.cs
@@ -1,4 +1,5 @@
 ﻿using Satori.AppServices.Services;
+using Satori.AppServices.ViewModels.AlertServices;
 
 namespace Satori.Pages.Components;
